requirements.txt
transformers
accelerate
datasets
faiss-cpu